Removed unused component
