--- src/pages/Admin/BasicInfo/BasicInfo.styles.ts
@@ -1,3 +1,0 @@
-import styled from "styled-components";
-
-export const Wrapper = styled.div``;
