--- src/pages/Admin/BasicInfo/BasicInfo.types.ts
@@ -1,6 +1,0 @@
-import { Venue_v2 } from "types/Venue";
-
-export interface BasicInfoProps {
-  venue: Venue_v2;
-  onSave: () => void;
-}
