--- src/pages/Admin/BasicInfo/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./BasicInfo";
